Hacktastic way of producing xml for the other content types.
This is not pretty, not pretty at all.  hand constructing xml goes against just about everything I've done
professionally for the last few years.  But I'm not keen to wade into the python xml soup  yet,
and this is all Proof Of Concept code, to see if we can get it working, not the final production code.