Junk the geomodel code, it's convenient, but far far too cpu costly.
Replaced with precalculated longitude buckets, at bar insert time, and then searching for all bars in a 1 degree wide
latitude band within that longitude bucket.  If we start to have performance problems, we can simply add more
buckets... ? This appears vastly faster, at least locally.

To go with this, overhauled the way we do bulk load of the osm files.  Checking for an existing bar by osm id was far
far far too slow as well, so as long as we made sure that we used the osm id in our key, we could simply _replace_ the
existing one with the new data, no matter what...

TODO - what impact does that have on pricings for a bar?